add new class at controller do transfer money and new endpoint at app

+++ controller/user.py
@@ -1,7 +1,7 @@
 from flask_restful import Resource, reqparse
 from models.user import UserModel
 from models.wallet import WalletModel
-from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
+from flask_jwt_extended import create_access_token, jwt_required
 from controller.security.safe_str_cmp import safe_str_cmp
 
 
@@ -11,6 +11,7 @@ atributos.add_argument('password', type=str, required=True, help=("The field 'ke
 atributos.add_argument('name', type=str, required=False)
 atributos.add_argument('cpf', type=str, required=False)
 atributos.add_argument('type', type=str, required=False)
+
 
 
 class User(Resource):
@@ -25,8 +26,11 @@ class User(Resource):
     @jwt_required()
     def delete(self, user_id):
         user = UserModel.find_user(user_id)
-        if user:
+        wallet = WalletModel.find_wallet(wallet_id=user_id)
+        if user and wallet:
             user.delete_user()
+            wallet.delete_wallet()
+            
             return{'message': 'User deleted'}, 201
         return {'message': 'User not found'}, 404
 
@@ -46,10 +50,11 @@ class UserRegister(Resource):
         wallet.save_wallet()
         return {"message": "User create successfully!"}, 201
 
+
 class UserLogin(Resource):      
         @classmethod
         def post(cls):
-            dados= atributos.parse_args()
+            dados = atributos.parse_args()
 
             user = UserModel.find_by_login(dados['email'])
 
@@ -57,3 +62,18 @@ class UserLogin(Resource):
                 acess_token = create_access_token(identity=user.user_id)
                 return {'acess_token': acess_token}, 200
             return {'message': 'The username or password is incorrect.'}, 401
+        
+
+class UserTransferMoney(Resource):
+        
+        @classmethod
+        def post(cls):
+            dados = atributos.parse_args()
+            wallet = WalletModel.find_wallet_by_cpf(dados['cpf'])
+
+            if wallet:
+                 return{'message': 'Carteira encontrada com sucesso'}
+
+            return{'message': 'Não foi possível encontrar a carteira do email {cpf}'.format(cpf=dados['cpf'])} 
+
+
